+++ Sources/iOS/TransitionController.swift
@@ -31,7 +31,7 @@
 import UIKit
 import Motion
 
-open class TransitionController: UIViewController {
+open class TransitionController: ViewController {
   /**
    A Boolean property used to enable and disable interactivity
    with the rootViewController.
@@ -100,11 +100,6 @@ open class TransitionController: UIViewController {
     return false
   }
   
-  open override func viewDidLoad() {
-    super.viewDidLoad()
-    prepare()
-  }
-  
   open override func viewWillAppear(_ animated: Bool) {
     super.viewWillAppear(animated)
     rootViewController.beginAppearanceTransition(true, animated: animated)
@@ -123,11 +118,6 @@ open class TransitionController: UIViewController {
   open override func viewDidDisappear(_ animated: Bool) {
     super.viewDidDisappear(animated)
     rootViewController.endAppearanceTransition()
-  }
-  
-  open override func viewWillLayoutSubviews() {
-    super.viewWillLayoutSubviews()
-    layoutSubviews()
   }
   
   /**
@@ -163,24 +153,8 @@ open class TransitionController: UIViewController {
     }
   }
   
-  /**
-   To execute in the order of the layout chain, override this
-   method. `layoutSubviews` should be called immediately, unless you
-   have a certain need.
-   */
-  open func layoutSubviews() {}
-  
-  /**
-   Prepares the view instance when intialized. When subclassing,
-   it is recommended to override the prepare method
-   to initialize property values and other setup operations.
-   The super.prepare method should always be called immediately
-   when subclassing.
-   */
-  open func prepare() {
-    view.clipsToBounds = true
-    view.backgroundColor = .white
-    view.contentScaleFactor = Screen.scale
+  open override func prepare() {
+    super.prepare()
     
     prepareContainer()
